docs: add PieChartWithPaddingAngle example in TypeScript
