+++ src/docs/exampleComponents/PieChart/PieChartWithPaddingAngle.tsx
@@ -1,0 +1,46 @@
+import { Cell, Pie, PieChart } from 'recharts';
+
+const data = [
+  { name: 'Group A', value: 400 },
+  { name: 'Group B', value: 300 },
+  { name: 'Group C', value: 300 },
+  { name: 'Group D', value: 200 },
+];
+const COLORS = ['#0088FE', '#00C49F', '#FFBB28', '#FF8042'];
+
+export default function Example() {
+  return (
+    <PieChart width={800} height={400}>
+      <Pie
+        data={data}
+        cx={120}
+        cy={200}
+        innerRadius={60}
+        outerRadius={80}
+        fill="#8884d8"
+        paddingAngle={5}
+        dataKey="value"
+      >
+        {data.map((entry, index) => (
+          <Cell key={`cell-${entry.name}`} fill={COLORS[index % COLORS.length]} />
+        ))}
+      </Pie>
+      <Pie
+        data={data}
+        cx={420}
+        cy={200}
+        startAngle={180}
+        endAngle={0}
+        innerRadius={60}
+        outerRadius={80}
+        fill="#8884d8"
+        paddingAngle={5}
+        dataKey="value"
+      >
+        {data.map((entry, index) => (
+          <Cell key={`cell-${entry.name}`} fill={COLORS[index % COLORS.length]} />
+        ))}
+      </Pie>
+    </PieChart>
+  );
+}
